fix(web): restore App component in Next.js
