+++ README.md
@@ -1,6 +1,6 @@
 # Too Many Types
 
-A minimal & pragmatic local-first ERP foundation built with **Fastify + PostgreSQL + Prisma** (API) and **Vite + React** (Web).
+A minimal & pragmatic local-first ERP foundation built with **Fastify + PostgreSQL + Prisma** (API) and **Next.js + React** (Web).
 
 ---
 
@@ -10,7 +10,7 @@ A minimal & pragmatic local-first ERP foundation built with **Fastify + PostgreS
 - **Roles**: `admin`, `user` (seeded at API start)
 - **API**: Fastify v5 + Zod validation, modular routes
 - **DB**: PostgreSQL + Prisma (schema versioned with migrations)
-- **Web**: Vite + React (TS) minimal UI for login/register
+- **Web**: Next.js + React (TS) minimal UI for login/register
 - **Modules**: Auto-load Fastify plugins placed under `packages/` (see `packages/example-module`)
 
 ---
@@ -19,7 +19,7 @@ A minimal & pragmatic local-first ERP foundation built with **Fastify + PostgreS
 
 - **Backend**: Fastify v5, @fastify/jwt, @fastify/cors, Prisma
 - **Database**: PostgreSQL 16+
-- **Frontend**: Vite + React (TypeScript)
+- **Frontend**: Next.js (pages router) + React (TypeScript)
 - **Tooling**: TypeScript, dotenv-cli, GitHub Actions
 - **Monorepo**: npm workspaces
 
@@ -31,7 +31,7 @@ A minimal & pragmatic local-first ERP foundation built with **Fastify + PostgreS
 too-many-types/
 ├─ apps/
 │  ├─ api/                 # Fastify + Prisma (schema in apps/api/prisma)
-│  └─ web/                 # Vite + React
+│  └─ web/                 # Next.js + React
 ├─ packages/
 │  ├─ example-module/      # sample Fastify plugin module
 │  └─ shared/              # (future) shared types/schemas/ui
@@ -144,7 +144,7 @@ Test with your JWT: `GET http://localhost:4000/admin/only` should return 200 for
 
 - Workflow: `.github/workflows/ci.yml`
 - **API job**: starts a Postgres 16 service, runs Prisma generate + migrate deploy, builds API
-- **Web job**: installs & builds the Vite app
+- **Web job**: installs & builds the Next.js app
 - Use env vars in CI instead of `.env` files
 
 ---
