+++ apps/web/styles/globals.css
@@ -1,0 +1,1 @@
+/* global styles */
